Adds telemetry
Adds span style events for pull request and ack, including tests and documentation.

Happy to take suggestions on naming. Potentially the existing demand is not something we should emit here? Added it cause it'd find it personally useful but happy to remove if it doesn't make sense.

It was a bit hard to do this because the config passed to ack doesn't come from the same place at all as the config passed to receive_messages, the first being just validated producer opts put in persistent term and not as you'd expect, the ack opts validated in the configure callback). And the second being the full config passed to the producer, including the broadway config.